requirements.txt
# VoiceInk - 核心依赖
sounddevice>=0.4.6
numpy>=1.24
pyyaml>=6.0

# ASR 引擎 (至少安装一个)
pywhispercpp>=0.2.0
# faster-whisper>=1.0.0    # 可选: pip install faster-whisper
funasr-onnx>=0.4.0          # Paraformer ONNX ASR 引擎 (中文精度更高)
jieba>=0.42                  # funasr-onnx 传递依赖
modelscope>=1.10             # ModelScope 模型下载

# LLM 推理
llama-cpp-python>=0.2.80
openai>=1.0                  # 云端 API 客户端 (可选)

# UI & 系统交互
pystray>=0.19
Pillow>=10.0
keyboard>=0.13
pyperclip>=1.8

# 工具
pypinyin>=0.49

# VAD
webrtcvad>=2.0.10
